update performance tests to match that of java's
